Reorganizing a library package
