+++ lib/src/month_item.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+/// Creates a Widget to represent the months.
+class MonthItem extends StatelessWidget {
+  final String name;
+  final Function onTap;
+  final bool isSelected;
+  final Color color;
+
+  MonthItem({this.name, this.onTap, this.isSelected, this.color});
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: this.onTap,
+      child: Text(
+        this.name.toUpperCase(),
+        style: TextStyle(
+          fontSize: 14,
+          color: color ?? Colors.black87,
+          fontWeight: this.isSelected ? FontWeight.bold : FontWeight.w300,
+        ),
+      ),
+    );
+  }
+}
