+++ lib/src/util/string_extension.dart
@@ -1,0 +1,8 @@
+extension StringExtension on String {
+  String capitalize() {
+    if (this.isEmpty) {
+      return this;
+    }
+    return this[0].toUpperCase() + this.substring(1);
+  }
+}
